pyproject.toml
[tool.poetry]
name = "instagram-daily-meme"
version = "0.1.0"
description = ""
authors = ["Marcuth <thiago.playgames15@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.7.1"
python-decouple = "^3.8"
selenium = "^4.20.0"
webdriver-manager = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
